Add frontend pdf generation (#39)
* Fix bug in overview

Technical and Fun skills were not shown in the overview as isTech and
isFun were not defined.

* Add slider and improve radio buttons layout

This adds a radio button for skill category (closes #31), and improves
radio buttons layout (closes #16).

* Fix dependency issues for paper-datatable

paper-datatable is not well
maintained (https://github.com/David-Mulder/paper-datatable/issues/112)

For now we regress to earlier versions of our packages.

* Add pdf generation

This introduces the fundamentals for generating a pdf (Re: #23).

Pressing on the pdf icon and then on the done icon shall generate the
pdf of the current content. Editing the content changes the pdf as well.