Routing on /cvs (#46)
* Fix bug in overview

Technical and Fun skills were not shown in the overview as isTech and
isFun were not defined.

* Add slider and improve radio buttons layout

This adds a radio button for skill category (closes #31), and improves
radio buttons layout (closes #16).

* Fix dependency issues for paper-datatable

paper-datatable is not well
maintained (https://github.com/David-Mulder/paper-datatable/issues/112)

For now we regress to earlier versions of our packages.

* Add pdf generation

This introduces the fundamentals for generating a pdf (Re: #23).

Pressing on the pdf icon and then on the done icon shall generate the
pdf of the current content. Editing the content changes the pdf as well.

* Fix dependencies issue again

The previous fix worked yesterday, but rebuilding the application made
the radio buttons unusable again. This change made the radio buttons
usable.

I am not sure this is fixed once for all though.

* Add swagger.yml documentation

* Move save button to cv-editor component

Re #10

* Fix typo

Re #19

* Make the routing work through cvs/username

This mostly completes #10. This commit has an issue with data binding:
to load data one has to pass through the cv list.

* Fix routing, fix minor bugs

This commit closes #10
Now everything works as a subroute of the cvs page.
The commit also solves minor issues with images.

* Fix photo input

Re #10

As we are using urls for images, make sense to use an input for now.

* Remove unused file

* Remove loading template

* Use the email as id in the address bar

With this commit the address bar will show
/cvs/andrea.giugliano@lunatech.com. However, it is not possible
anymore to reload the page, as of Polymer's Poyserve issue 147.

* Use dashed urls

For the Polymer Polyserve issue we need to use dashed url: so for
/cvs/andrea.giugliano@lunatech.com we will use
/cvs/andrea-giugliano@lunatech-com.